+++ src/routes/api/pusher/auth_user/+server.ts
@@ -2,11 +2,18 @@ import { json } from '@sveltejs/kit'
 import type { RequestHandler } from './$types'
 import { pusher } from '$lib/utils/pusher'
 
+// POST:	Authenticate user
 export const POST: RequestHandler = async ({ request }) => {
 	const body = await request.text()
 	const socketId = body.match(/socket_id=([^&]*)/)?.[1] ?? ''
-	const channel = body.match(/channel_name=([^&]*)/)?.[1] ?? ''
 
-	const authResponse = pusher.authorizeChannel(socketId, channel)
+	const user = {
+		id: '0000-0000',
+		user_info: {
+			name: 'Nosson M Frankel'
+		}
+	}
+
+	const authResponse = pusher.authenticateUser(socketId, user)
 	return json(authResponse)
 }