+++ src/routes/api/pusher/send/+server.ts
@@ -1,18 +1,12 @@
 import { json } from '@sveltejs/kit'
 import type { RequestHandler } from './$types'
-import Pusher from 'pusher'
+import { pusher } from '$lib/utils/pusher'
 
 // GET:    Test pusher implemention
-export const GET: RequestHandler = () => {
-	const pusher = new Pusher({
-		appId: import.meta.env.VITE_PUSHER_APP_ID,
-		key: import.meta.env.VITE_PUSHER_KEY,
-		secret: import.meta.env.VITE_PUSHER_SECRET,
-		cluster: 'eu',
-		useTLS: true
-	})
+export const GET: RequestHandler = ({ url }) => {
+	const userId = url.searchParams.get('userId')
 
-	pusher.trigger('private-chat-userId', 'message', {
+	pusher.trigger(`private-chat-${userId}`, 'msg', {
 		message: 'hello world'
 	})