+++ src/routes/pusher/+page.svelte
@@ -9,8 +9,8 @@
 		var pusher = new Pusher(import.meta.env.VITE_PUSHER_KEY, {
 			cluster: 'eu',
 			// userAuthentication: {
-				// endpoint: "/api/pusher/auth_channel",
-				// transport: "ajax",
+			// endpoint: "/api/pusher/auth_channel",
+			// transport: "ajax",
 			// },
 			channelAuthorization: {
 				endpoint: '/api/pusher/auth_channel',
@@ -20,11 +20,6 @@
 
 		pusher.signin()
 
-		// const channel = pusher.subscribe('chat');
-		// channel.bind('message', data => {
-		// 	messages = [...messages, data];
-		// });
-		// var channel = pusher.subscribe('my-channel')
 		var channel = pusher.subscribe('private-chat-userId')
 		channel.bind('message', (data) => {
 			alert(JSON.stringify(data))